ENA_1/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
arclet-alconna==1.8.36
arclet-alconna-tools==0.7.10
cachetools==5.5.2
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
cookit==0.10.1
distro==1.9.0
emoji==2.14.1
exceptiongroup==1.2.2
fastapi==0.115.12
greenlet==3.1.1
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.6
jiter==0.9.0
loguru==0.7.3
Markdown==3.7
MarkupSafe==3.0.2
msgpack==1.1.0
multidict==6.2.0
nepattern==0.7.7
nonebot-adapter-onebot==2.4.6
nonebot-plugin-alconna==0.56.2
nonebot-plugin-apscheduler==0.5.0
nonebot-plugin-htmlrender==0.6.2
nonebot-plugin-picstatus==2.1.3.post1
nonebot-plugin-waiter==0.8.1
nonebot2==2.4.2
nonebot_plugin_userinfo==0.2.6
openai==1.76.0
pillow==11.1.0
playwright==1.51.0
propcache==0.3.1
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic==2.11.1
pydantic_core==2.33.0
pyee==12.1.1
Pygments==2.19.1
pygtrie==2.5.0
pymdown-extensions==10.14.3
python-dotenv==1.1.0
python-markdown-math==0.8
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.46.1
StrEnum==0.4.15
tarina==0.6.8
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
tzlocal==5.3.1
uvicorn==0.34.0
watchfiles==1.0.4
websockets==15.0.1
win32_setctime==1.2.0
yarl==1.18.3
zipp==3.21.0
